requirements.txt
requests>=2.31.0
yfinance>=0.2.36
reportlab>=4.1.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.3
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
slowapi>=0.1.9
aiosqlite>=0.19.0
